Add validations for the ID's to delete a specific workout

+++ back-end/controllers/workoutController.js
@@ -42,11 +42,11 @@ const createWorkout = async (req, res) => {
 const deleteWorkout = async (req, res) => {
   const { id } = req.params;
 
-  const workout = await Workout.findByIdAndDelete(id);
-
-  if (!workout) {
-    return res.status(404).json({ error: "No such workout" });
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({ error: "No such workout(Invalid Id)" });
   }
+
+  const workout = await Workout.findByIdAndDelete(id);
 
   res.status(200).json({ message: "Workout deleted successfully" });
 };
